fix: Fix unwanted automatic navigation back to QuizResults
Resetting questionsAnswered to 0 is necessary to avoid a bug where the
app would navigate back to QuizResults after arriving at the Deck view.
It appears to be an issue where the Quiz view's componentDidUpdate is
being called as the app navigates back through the navigation stack. In
that call, questionsAnswered is compared to the deck length. If the
values are equal, the app navigates to the QuizResults view. By setting
the questionsAnswered to 0, this comparison returns false and the app
does not navigate back to QuizResults on arrival to Deck.