fix: Move resettings of quiz to didMount for Deck
The bug here was that taking a quiz with 2 questions for example would
have the currentQuestion at 2 on quiz completion. When a quiz would be
attempted in a deck with 1 question, there would be an initial render
attempt of the quiz where the resetQuiz method called on
componentDidMount for Quiz would not have reset the currentQuestion yet.
So with the currentQuestion set at 2, the currentCard would be set with
an index that doesn't exist in the deck. This resulted in error of
trying to access the id of an undefined currentCard. Moving the
resetQuiz to before the Quiz is navigated to eliminates this problem.